Reestructured the model of the assistant application.
