+++ assistant/models/assistant_document.py
@@ -1,0 +1,38 @@
+##############################################################################
+#
+#    OSIS stands for Open Student Information System. It's an application
+#    designed to manage the core business of higher education institutions,
+#    such as universities, faculties, institutes and professional schools.
+#    The core business involves the administration of students, teachers,
+#    courses, programs and so on.
+#
+#    Copyright (C) 2015-2016 Université catholique de Louvain (http://www.uclouvain.be)
+#
+#    This program is free software: you can redistribute it and/or modify
+#    it under the terms of the GNU General Public License as published by
+#    the Free Software Foundation, either version 3 of the License, or
+#    (at your option) any later version.
+#
+#    This program is distributed in the hope that it will be useful,
+#    but WITHOUT ANY WARRANTY; without even the implied warranty of
+#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#    GNU General Public License for more details.
+#
+#    A copy of this license - GNU General Public License - is available
+#    at the root of the source code of this program.  If not,
+#    see http://www.gnu.org/licenses/.
+#
+##############################################################################
+from django.db import models
+from django.utils.translation import ugettext_lazy as _
+
+
+class AssistantDocument(models.Model):
+    DOC_TYPE_CHOICES = (
+        ('PHD', _('PhD')),
+        ('TUTORING', _('Tutoring')),
+        ('RESEARCH', _('Research')))
+
+    assistant = models.ForeignKey('AcademicAssistant')
+    mandate = models.ForeignKey('AssistantMandate')
+    doc_type = models.CharField(max_length=20, choices=DOC_TYPE_CHOICES)
